refactor(server): move org API handlers out of handlers/admin
Move the five /api/v1/org/* JSON handlers (SCIM-token create/list/
delete, audit-events export, policy validation) plus their private
ocsf helper module from handlers/admin/ into a new handlers/api/org/
module, so the file layout matches the routing: handlers/admin/ now
holds only browser pages and form POSTs, and the JWT-authenticated
JSON API surface lives under handlers/api/.

Shared org-admin helpers stay shared in handlers/admin/mod.rs as
pub(crate): generate_scim_token, compute_token_expiry, plus the
hoisted MAX_SCIM_TOKEN_DESCRIPTION_CHARS, requested_scope, and
has_audit_read. Module docs describe each handler's actual auth
shape (Bearer/DPoP with session-cookie fallback; audit_events
additionally accepts an org API token with audit:read and rejects
requests without an Authorization header).

No behavior change: same routes, extractors, responses, and
rate-limit groups. Tests move with their handlers unmodified apart
from module paths; the setup_admin fixture is promoted verbatim to
test_utils::create_test_org_admin for use by both sides of the
split.

Closes #1176
